Updated support for integer division behavior
This adds an adapted fix to give conditional integer division support,
based on the environment variable MRB_INTEGER_DIVISION. This adapts from
the previous fix to the latest stable mRuby branch, taking into
account their new MRB_WITHOUT_FLOAT environment variable.

Note that MRB_WITHOUT_FLOAT and MRB_INTEGER_DIVISION are not compatible
and should not be used together.

With the former unset and the latter set, we can build an mRuby that
supports native floating point values, but still gives an integer
quotient result for dividing two integers (non-integer dividend or
divisor give float results, as usual).

This fix is adapted from the fix in this open PR on the mRuby project:
https://github.com/mruby/mruby/pull/3123